requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
cryptography==44.0.3
cssselect2==0.8.0
daphne==4.1.2
defusedxml==0.7.1
Django==5.2
django-allauth==65.7.0
django-celery==3.1.17
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-paypal==2.1
django-recaptcha==3.0.0
django-redis==5.4.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
lxml==5.4.0
msgpack==1.1.0
oscrypto==1.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pyOpenSSL==25.0.0
pypdf==5.4.0
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==5.3.0
reportlab==4.4.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
svglib==1.5.1
tinycss2==1.4.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==7.2
